move connect hoc to provider
